pyproject.toml
[project]
name = "perfectdeckcli"
version = "0.1.0"
description = "Hybrid CLI + MCP tool to manage App Store and Play Store listing data."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "mcp>=1.13.0",
  "pydantic>=2.8.0",
  "pyyaml>=6.0.1",
  "google-api-python-client>=2.0.0",
  "google-auth>=2.0.0",
  "PyJWT>=2.0.0",
  "cryptography>=3.0.0",
  "requests>=2.28.0",
  "jq>=1.8.0",
]

[project.scripts]
perfectdeckcli = "perfectdeckcli.cli:main"
perfectdeck-mcp = "perfectdeckcli.mcp_server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
